tyytväisyysprosentti kokonaisluvuksi

+++ Simulator/src/controller/PaneelitController.java
@@ -205,7 +205,7 @@ public class PaneelitController implements IKontrolleriForV, IKontrolleriForM {
 	public void updateTyytyvaisyys(double v) {
 		Platform.runLater(new Runnable() {
 			public void run() {
-				tyytyvProsLuku.setText(String.valueOf(v));
+				tyytyvProsLuku.setText(String.valueOf((int)v));
 			}
 		});
